brcmfmac_sdio-firmware: Select correct firmware for BananaPi M2 Zero
The BananaPi M2 Zero comes with an AP6212 broadcom wifi and bluetooth
combo chip. Apparently, devices differ in terms of which bluetooth
firmware the should load, even if the kernel selects one and the
same file. Make sure that the right AP6212 bluetooth firmware for
BananaPi M2 Zero is selected.